Resolves issue #94
min and max values must be doubles (not int), allow colors to be set
equal to min and max values

+++ LinkedView/src/Controllers/TVController.java
@@ -699,8 +699,8 @@ public class TVController {
 				
 				if(menu.equalsIgnoreCase(StringRes.menu_Color)) {
 					
-					int min = (int)model.getDataMatrix().getMinVal();
-					int max = (int)model.getDataMatrix().getMaxVal();
+					Double min = model.getDataMatrix().getMinVal();
+					Double max = model.getDataMatrix().getMaxVal();
 					
 					/* View */
 					ColorChooser gradientPick =
